+++ frontend/index.html
@@ -1,38 +1,35 @@
 <!doctype html>
 <html lang="en">
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-<head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <meta name="Author" content="" />
+        <meta name="Developer" content="" />
+        <meta name="Owner" content="" />
 
-    <meta name="Author" content="">
-    <meta name="Developer" content="">
-    <meta name="Owner" content="">
+        <title>Elephoria</title>
 
-    <title>Elephoria</title>
+        <!-- Favicon Icons -->
+        <link href="apple-touch-icon.png" rel="apple-touch-icon" sizes="180x180" />
+        <link type="image/png" href="favicon-32x32.png" rel="icon" sizes="32x32" />
+        <link type="image/png" href="favicon-16x16.png" rel="icon" sizes="16x16" />
+        <link href="site.webmanifest" rel="manifest" />
+        <link href="safari-pinned-tab.svg" rel="mask-icon" color="#FFB6B9" />
+        <meta name="msapplication-TileColor" content="#FFB6B9" />
+        <meta name="theme-color" content="#FFB6B9" />
+        <meta name="dark-theme-color" content="#1f2937" />
 
-    <!-- Favicon Icons -->
-    <link href="apple-touch-icon.png" rel="apple-touch-icon" sizes="180x180">
-    <link type="image/png" href="favicon-32x32.png" rel="icon" sizes="32x32">
-    <link type="image/png" href="favicon-16x16.png" rel="icon" sizes="16x16">
-    <link href="site.webmanifest" rel="manifest">
-    <link href="safari-pinned-tab.svg" rel="mask-icon" color="#FFB6B9">
-    <meta name="msapplication-TileColor" content="#FFB6B9">
-    <meta name="theme-color" content="#FFB6B9">
-    <meta name="dark-theme-color" content="#1f2937">
+        <!-- Preconnect to Google Fonts -->
+        <link rel="preconnect" href="https://fonts.googleapis.com" />
+        <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
 
+        <!-- Google Fonts -->
+        <link href="https://fonts.googleapis.com/css2?family=Aladin&family=Indie+Flower&display=swap" rel="stylesheet" />
+    </head>
 
-    <!-- Preconnect to Google Fonts -->
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-
-    <!-- Google Fonts -->
-    <link href="https://fonts.googleapis.com/css2?family=Aladin&family=Indie+Flower&display=swap" rel="stylesheet">
-</head>
-
-<body class="bg-primary">
-    <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
-</body>
-
+    <body class="bg-primary">
+        <div id="root"></div>
+        <script type="module" src="/src/main.jsx"></script>
+    </body>
 </html>
