+++ frontend/src/assets/svg/app-name.jsx
