+++ frontend/src/assets/svg/close.jsx
