+++ frontend/src/assets/svg/gamepad.jsx
