+++ frontend/src/assets/svg/lock.jsx
