+++ frontend/src/assets/svg/logo.jsx
