+++ frontend/src/assets/svg/players.jsx
