+++ frontend/src/assets/svg/reset.jsx
