+++ frontend/src/components/PrivateRoute.jsx
@@ -1,15 +1,17 @@
-import React from 'react';
-import { Navigate } from 'react-router-dom';
-import { useAuth } from '../hooks/useAuth';
+import React from 'react'
+
+import { Navigate } from 'react-router-dom'
+
+import { useAuth } from '../hooks/useAuth'
 
 const PrivateRoute = ({ children }) => {
-  const { user, loading } = useAuth();
+    const { user, loading } = useAuth()
 
-  if (loading) {
-    return <div>Loading...</div>;
-  }
+    if (loading) {
+        return <div>Loading...</div>
+    }
 
-  return user ? children : <Navigate to="/auth" />;
-};
+    return user ? children : <Navigate to="/auth" />
+}
 
-export default PrivateRoute;
+export default PrivateRoute
