+++ frontend/src/hooks/useAuth.js
@@ -1,19 +1,20 @@
-import { useEffect, useState } from 'react';
-import { getAuth, onAuthStateChanged } from 'firebase/auth';
+import { useEffect, useState } from 'react'
+
+import { getAuth, onAuthStateChanged } from 'firebase/auth'
 
 export const useAuth = () => {
-    const [user, setUser] = useState(null);
-    const [loading, setLoading] = useState(true);
+    const [user, setUser] = useState(null)
+    const [loading, setLoading] = useState(true)
 
     useEffect(() => {
-        const auth = getAuth();
+        const auth = getAuth()
         const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
-            setUser(currentUser);
-            setLoading(false);
-        });
+            setUser(currentUser)
+            setLoading(false)
+        })
 
-        return () => unsubscribe();
-    }, []);
+        return () => unsubscribe()
+    }, [])
 
-    return { user, loading };
-};
+    return { user, loading }
+}
