+++ frontend/src/main.jsx
@@ -1,4 +1,5 @@
 import { StrictMode } from 'react'
+
 import { createRoot } from 'react-dom/client'
 
 import App from './App'
