+++ frontend/src/pages/games/tic-tac-toe/components/heading.jsx
@@ -1,4 +1,5 @@
 import React from 'react'
+
 import { Link } from 'react-router-dom'
 
 import NeuButton from '../../../../components/common/buttons/neu-button'
