+++ frontend/src/pages/games/tic-tac-toe/components/player-name-modal.jsx
@@ -2,8 +2,8 @@ import React, { useEffect, useRef, useState } from 'react'
 
 import { Icon } from '@iconify/react'
 
+import Players from '../../../../assets/svg/players'
 import NeuButton from '../../../../components/common/buttons/neu-button'
-import Players from '../../../../components/common/svg/players'
 
 const PlayerNameModal = ({ setPlayerName, playerXName, playerOName, closeModal }) => {
     const [error, setError] = useState(false)
